requirements.txt
ffmpeg-python==0.2.0
binaryornot==0.4.4